debug E2E test

+++ scripts/tests/deploy_moduletemplate.sh
@@ -8,6 +8,7 @@ MODULE_NAME=$1
 RELEASE_VERSION=$2
 INCLUDE_DEFAULT_CR=${3:-true}
 MANDATORY=${4:-false}
+DEPLOY_MODULETEMPLATE=${5:-true}
 
 cat <<EOF > module-config-for-e2e.yaml
 name: kyma-project.io/module/${MODULE_NAME}
@@ -36,13 +37,16 @@ fi
 cat module-config-for-e2e.yaml
 modulectl create --config-file ./module-config-for-e2e.yaml --registry http://localhost:5111 --insecure
 sed -i 's/localhost:5111/k3d-kcp-registry.localhost:5000/g' ./template.yaml
-kubectl apply -f template.yaml
 
 cat template.yaml
 echo "ModuleTemplate created successfully"
 
+if [ "${DEPLOY_MODULETEMPLATE}" == "true" ]; then
+kubectl apply -f template.yaml
+rm -f template.yaml
+fi
+
 rm -f module-config-for-e2e.yaml
 rm -f template-operator.yaml
-rm -f template.yaml
 rm -f default-sample-cr.yaml
 echo "Temporary files removed successfully"
